requirements.txt
numpy==2.2.2
tiktoken==0.8.0
backoff==2.2.1
anthropic==0.44.0
tqdm==4.67.1
openai==1.59.9
pyyaml==6.0.2
attrs==24.3.0